Refactor PPL SLA report
The old report was including applications which were originally submitted as not complete and correct because the original submission was always setting the submitted timestamp.

Refactor to only set the submission timestamp when the application is complete and correct.

Additionally cover some edge cases so the data output makes sense if a deadline passes and then an application is returned and resubmitted (where peviously the submission date would update to after the deadline). Now the submission date in the report is always the first submission which passed a deadline, and is not updated by resubmission.