Include times from resubmission to action in tasks report
Where a subsequent submission event is detected set a pointer to the resubmission date and then when actions are detected push the time since resubmission to an array of resubmission times.

This will allow the data exports builder to create a mean time from resubmission to action for tasks by type as well as the first time submission.

Also move the `returnCount` increment out of the condition so that it actually represents the number of returns, and not just a binary flag of whether the task was ever returned.